Move key modules into flit_core
